# Upgrade to Webpack 4