--- a/package.json
+++ b/package.json
@@ -3,9 +3,6 @@
   "version": "4.46.0",
   "description": "An item manager for Destiny.",
   "main": "app/index.html",
-  "engines": {
-    "node": "6.9.2"
-  },
   "scripts": {
     "test": "echo \"No tests\"",
     "lint": "grunt lintJSON && htmlhint && eslint --fix . && tslint --fix -p tsconfig.json",