--- a/src/app/d2-vendors/vendor.scss
+++ b/src/app/d2-vendors/vendor.scss
@@ -1,5 +1,6 @@
 // This inherits a lot of styles from ../vendors/vendors.scss
 @import '../../scss/variables.scss';
+@import '../vendors/vendors.scss';
 
 .d2-vendors {
   padding: 0 1em;