--- a/src/return.html
+++ b/src/return.html
@@ -1,10 +1,6 @@
 <html !doctype>
   <head>
     <title>Validating Authorization</title>
-
-    {{#each htmlWebpackPlugin.files.css}}
-    <link href="{{ webpackConfig.output.publicPath }}{{ this }}" rel="stylesheet">
-    {{/each}}
   </head>
 
   <body>
@@ -17,10 +13,5 @@
         <a href="/index.html">Go back to DIM and try again</a>
       </div>
     </div>
-
-    {{{htmlWebpackPlugin.files.webpackManifest}}}
-    {{#each htmlWebpackPlugin.files.js}}
-    <script type="text/javascript" src="{{ webpackConfig.output.publicPath }}{{ this }}"></script>
-    {{/each}}
   </body>
 </html>