This upgrades to Webpack 4. There isn't a huge benefit - it's a bit faster, and maybe a little smaller output. It fixes a couple annoying bugs and allows for more future optimization through code splitting though.