@SunburnedGoose 2018-04-13
Nice. I tried to do this and didn't get too far. Would be good to learn the before and after.