@bhollis 2018-04-15
Before:

```
before webpack 4

-rw-r--r--+ 1 brh  staff    45973 Apr  8 17:08 dist/authReturn-624715.js
-rw-r--r--+ 1 brh  staff    37258 Apr  8 17:08 dist/browsercheck-dea136.js
-rw-r--r--+ 1 brh  staff    68422 Apr  8 17:08 dist/chunk-0-destiny1-ac4533.js
-rw-r--r--+ 1 brh  staff  1222450 Apr  8 17:08 dist/chunk-1-sqlLib-73b6a0.js
-rw-r--r--+ 1 brh  staff      742 Apr  8 17:08 dist/gdriveReturn-48f9d5.js
-rw-r--r--+ 1 brh  staff    58000 Apr  8 17:08 dist/inflate-fd952b.js
-rw-r--r--+ 1 brh  staff   482017 Apr  8 17:08 dist/main-0078fc.js
-rw-r--r--+ 1 brh  staff     1433 Apr  8 17:08 dist/manifest-c08dd1.js
-rw-r--r--+ 1 brh  staff    55729 Apr  8 17:08 dist/service-worker.js
-rw-r--r--+ 1 brh  staff    61492 Apr  8 17:08 dist/sql-wasm-0edda1.js
-rw-r--r--+ 1 brh  staff  1029945 Apr  8 17:08 dist/vendor-c0560f.js
-rw-r--r--+ 1 brh  staff     3679 Apr  8 17:08 dist/z-worker-7d40d9.js
```

After:

```
-rw-r--r--+ 1 brh  staff     3499 Apr  8 17:04 dist/authReturn-fafbf0.js
-rw-r--r--+ 1 brh  staff     1107 Apr  8 17:04 dist/browsercheck-8c47c7.js
-rw-r--r--+ 1 brh  staff    67410 Apr  8 17:04 dist/destiny1-4bf301.js
-rw-r--r--+ 1 brh  staff      807 Apr  8 17:04 dist/gdriveReturn-59e548.js
-rw-r--r--+ 1 brh  staff    58000 Apr  8 17:04 dist/inflate-fd952b.js
-rw-r--r--+ 1 brh  staff   476600 Apr  8 17:04 dist/main-1e9d6b.js
-rw-r--r--+ 1 brh  staff     2690 Apr  8 17:04 dist/runtime-cfdcc4.js
-rw-r--r--+ 1 brh  staff    46554 Apr  8 17:04 dist/service-worker.js
-rw-r--r--+ 1 brh  staff    61492 Apr  8 17:04 dist/sql-wasm-0edda1.js
-rw-r--r--+ 1 brh  staff    42730 Apr  8 17:04 dist/vendors-authReturn-main-75ce5b.js
-rw-r--r--+ 1 brh  staff    36319 Apr  8 17:04 dist/vendors-browsercheck-d2babf.js
-rw-r--r--+ 1 brh  staff   985755 Apr  8 17:04 dist/vendors-main-5e2309.js
-rw-r--r--+ 1 brh  staff  1222515 Apr  8 17:04 dist/vendors-sqlLib-25b426.js
-rw-r--r--+ 1 brh  staff     3679 Apr  8 17:04 dist/z-worker-7d40d9.js
```

I think the folks that were seeing big benefits were mostly on the defaults - those have gotten way better. We already had a really optimized config though.